Changed jmp to sjmp
Jesus says "Found the problem!  In your code you are using the
pseudo-instruction 'jmp'. The assembler tries to best guess what is the
best instruction to use instead: sjmp, ajmp, or ljmp.  Somewhere the
assembler gets confused and messes up. The work around to this problem
for now is to replace every 'jmp' in your code with either 'ljmp' or
'sjmp'."